requirements.txt
flask~=2.0
requests>=2.26
schedule>=1.1
pyyaml>=6.0
